refactor: entities folder renaming
