+++ src/domain/use_cases/student/getById.student.ts
@@ -1,6 +1,6 @@
 import { Inject, Injectable } from '@nestjs/common';
 import { Repository } from 'typeorm';
-import { Student } from '../../entitys/student.entity';
+import { Student } from '../../entities/student.entity';
 
 @Injectable()
 export class GetByIdStudent {
