+++ test/use-cases/student/update.student.spec.ts
@@ -1,6 +1,6 @@
 import { Connection, Repository, UpdateResult } from 'typeorm';
 import { UpdateStudentDTO } from '../../../src/domain/dto/student/updateStudent.dto';
-import { Student } from '../../../src/domain/entitys/student.entity';
+import { Student } from '../../../src/domain/entities/student.entity';
 import { UpdateStudent } from '../../../src/domain/use_cases/student';
 import { mockUpdateStudentDto } from '../../studentData';
 import { testsAppModule } from '../../test.app.module.factory';
